Make failure scripts way more resilient
Previously they could fail if a page interrupted the script (i.e.
quickly changing location.href after we do), or if we got notified
early, and started injecting into the page we're leaving rather than the
new page we're going to (especially if we're not allowed to, as with
chrome://newtab).